7 Code review suggestions
